CloudFormation Script to create network resources.
This template deploys a VPC, with a pair of public and private subnet across two availability zones. It deploys an Internet Gateway, with a default route on the public subnets. It deploys a pair of NAT Gateways(one in each AZ) and default routes for them in the private subnets.